Fix jet moment decoration handling

+++ Reconstruction/Jet/JetMomentTools/JetMomentTools/JetWidthTool.h
@@ -35,7 +35,7 @@ public:
   virtual StatusCode decorate(const xAOD::JetContainer& jets) const;
 
   // Local method to calculate and return the width.
-  double width(const xAOD::Jet& jet, double& widthEta, double& widthPhi) const;
+  float width(const xAOD::Jet& jet, float& widthEta, float& widthPhi) const;
 
 private:
